scons: Add cxx_config support for OptionalParams
The initial implementation of OptionalParams [1] does not support
cxx_config builds and therefore breaks some of our CI tests, as reported
by a recent PR [2].

The issue is trying to unserialize a std::optional<int> from a config
file (ini). The to_number does not allow a std::optional parameter.

While at the beginning I thought about specializing to_number,
I realised this is not generic enough as it covers the
std::optional<int/unsigned etc> only cases and it does not
cover when the parameter is not a number.

[1]: https://github.com/gem5/gem5/pull/2252
[2]: https://github.com/gem5/gem5/pull/2270

Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>
Change-Id: Ibb03e26b9390fbdc5b7a369a1c00d5ff625923ce