arch-arm: Generate disable PMU ExitEvent when local disable
With the ArmPMU it is possible to generate an exit event
whenever the PMU gets enabled/disabled. The obvious way
to check for it is to catch writes to the PMCR register
and generate the exit event accordingly.

While this works well for enabling events, we have seen
cases where userspace does not disable the PMU globally
at the end of the region of interest, and instead simply
disables all counters by writing 0xffff_ffff to the
PMCNTENCLR register.

We now cover this case by generating an exit event provided
all counters are disabled

Change-Id: I9c487f8dc8e423a3c86ac7bbd7af90b5d0fb9e4d
Signed-off-by: Giacomo Travaglini <giacomo.travaglini.arm.com>
Reviewed-by: Andreas Sandberg <andreas.sandberg@arm.com>